Adds necessary packages to revel/modules and pins revel version to the correct commit
The web app needed specific subpackages of the revel modules, which dep did not import by default. The redirect error of fsnotify is resolved by pinning specific commits of the v17 release branch(not tag), which contain the correct import path for fsnotify. The Makefile is also running revel, so the specific branches are used as well.