Selection between three track choices available (#22)
* added tags field to yaml file

* adding tag list to TaskEntry struct

* removed quotes

* tag selection logic in workload.go

* go formatted

* successfully adding logic to workload to select by track tag

* adding an extra env var to app.conf and making appropriate changes.

* first attempts to add a separate website for tracks

* added addTrack method to controller to handle form input to pass along. Added input form. Added changes to workload and firguring the ins and outs of routing.

* passing form entry into next view successfully

* adding Track to User model as well as GenerateProject struct and attempting to pass it around.

* can take in multiple args from check boxes and pass around a slice of tracks

* checking if tags are in clicked tracks to generate issues

* Workable user facing selection of multiple tracks

The user can select check boxes to choose from three tracks: cluster operator, cnct hire, and app developer. The onboarding app will generate or update a project with all the cards whose tags match any of the tracks chosen. For issues that exist in multiple tracks, currently an error message gets generated, which is a situation that will need further refinement; we may not want an error message to show up for a task that got generated during the same action, just because it appears in two tracks. Also needing improvement is the order in which the cards appear.

* Track view cleanup

* cleaning up unnecessary track var from env setup