Dynamically read, offer onboarding track choices and generate appropriate tasks from yaml file in web app view (#23)
* Added POST route for form

* Reading from form via key value. Tasks not rendering currently

* Adding form values to track var successfully. Page won't create the tasks though

* after changing User.Tracks it will generate tasks per track

* Removed GetTracks route as not needed

* Removed GetTracks controller method, pass and read in checkboxes dynamically in from. Also added TODOs for further improvement

* making the checklist labels make sense

* Added AvailableTracks as a field to the User Model and adjusted handling within controllers and views accordingly.

To read in the learning tracks available, a new field was added to the yaml file and read in in config.go. Remaining work could improve on this by parsing the yaml and reading the available tasks from a collection of tags from all the tasks. Additionally, error messages for re-rendering tasks that got created under a previous track need cleaning up.

Overall, a dynamic rendering of available tracks via a check box web from.

* Cleanup of debugging statements and adding comments to controller methods

* Adding track selection method to controller

app.Setup has all the information from the yaml file, including which tracks we tagged each task with. The new method scans eatch task's tags, determines the unique tracks and appends them to a slice of tracks that this particular yaml file provides.

This concludes the main body of dynamic selection and rendering of multiple tracks. Some cleanup remains, as well as a better descriptor for the user-facting checklist.

* Final cleanup

* Sample onboarding yaml file with tags

* streamlined code comments